+++ testSuite/unit/MsqMeshEntityTest.cpp
@@ -8,7 +8,7 @@
 //    E-MAIL: tleurent@mcs.anl.gov
 //
 // ORIG-DATE: 12-Nov-02 at 18:05:56
-//  LAST-MOD:  5-Dec-02 at 16:55:01 by Thomas Leurent
+//  LAST-MOD:  9-Apr-03 at 10:03:12 by Thomas Leurent
 //
 // DESCRIPTION:
 // ============
@@ -85,8 +85,8 @@ public:
     MsqError err;
     // prints out the vertices.
     MsqVertex* ideal_vertices = one_hex_patch.get_vertex_array(err); MSQ_CHKERR(err);
-    int num_vtx = one_hex_patch.num_vertices();
-    CPPUNIT_ASSERT_EQUAL(8, num_vtx);
+    size_t num_vtx = one_hex_patch.num_vertices();
+    CPPUNIT_ASSERT_EQUAL(size_t(8), num_vtx);
     
     MsqVertex vtx;
 
@@ -112,7 +112,7 @@ public:
     hex->get_sample_points(QualityMetric::ELEMENT_VERTICES, sample_points, err); MSQ_CHKERR(err);
       // and get the jacobian vectors (should be the identity matrix).
       //for the first sample point (0,0,0)
-    int num_jac_vec;
+    short num_jac_vec;
     std::vector<Vector3D>::iterator sp;
     sp=sample_points.begin();
     Vector3D jacobian_vectors[3];
@@ -149,7 +149,7 @@ public:
     std::vector<Vector3D> sample_points;
     tet->get_sample_points(QualityMetric::ELEMENT_VERTICES, sample_points, err); MSQ_CHKERR(err);
     // and get the jacobian vectors (should be the identity matrix).
-    int num_jac_vec;
+    short num_jac_vec;
     std::vector<Vector3D>::iterator sp;
     sp=sample_points.begin();
     Vector3D jacobian_vectors[3];
@@ -185,7 +185,7 @@ void test_compute_weigted_jacobian_ideal_quad()
     std::vector<Vector3D> sample_points;
     quad->get_sample_points(QualityMetric::ELEMENT_VERTICES, sample_points, err); MSQ_CHKERR(err);
     // and get the jacobian vectors (should be the identity matrix).
-    int num_jac_vec;
+    short num_jac_vec;
     std::vector<Vector3D>::iterator sp;
     sp=sample_points.begin();
     Vector3D jacobian_vectors[3];
@@ -220,7 +220,7 @@ void test_compute_weigted_jacobian_ideal_tri()
     std::vector<Vector3D> sample_points;
     tri->get_sample_points(QualityMetric::ELEMENT_VERTICES, sample_points, err); MSQ_CHKERR(err);
     // and get the jacobian vectors (should be the identity matrix).
-    int num_jac_vec;
+    short num_jac_vec;
     std::vector<Vector3D>::iterator sp;
     sp=sample_points.begin();
     Vector3D jacobian_vectors[3];
